change floating point type in SimConfig to float

+++ demo/src/config.hpp
@@ -10,12 +10,12 @@ namespace config {
 //*          Global             */
 //********************************/
 struct Global {
-  double dt = 0.01;
+  float dt = 0.01;
   int max_outer_iteration = 100;
   int max_inner_iteration = 50;
-  std::array<double, 3> acceleration{0.0, 0.0, -9.8};
+  std::array<float, 3> acceleration{0.0, 0.0, -9.8};
   int total_steps = 100;
-  double max_time = 0.0;
+  float max_time = 0.0;
   bool headless = false;
 };
 
@@ -26,27 +26,27 @@ struct Collision {
   bool enabled = true;
   bool self_collision = false;
   int group = 0;
-  double restitution = 0.0;
-  double friction = 0.5;
+  float restitution = 0.0;
+  float friction = 0.5;
 };
 
 //********************************/
 //*          Transform           */
 //********************************/
 struct Transform {
-  std::array<double, 3> translation{0.0, 0.0, 0.0};
-  std::array<double, 3> rotation_euler_deg{0.0, 0.0, 0.0};
-  std::array<double, 3> scale{1.0, 1.0, 1.0};
+  std::array<float, 3> translation{0.0, 0.0, 0.0};
+  std::array<float, 3> rotation_euler_deg{0.0, 0.0, 0.0};
+  std::array<float, 3> scale{1.0, 1.0, 1.0};
 };
 
 //********************************/
 //*         ClothParams          */
 //********************************/
 struct ClothParams {
-  double elastic_stiffness = 100.0;
-  double bending_stiffness = 0.0001;
-  double density = 0.1;
-  double damping = 0.01;
+  float elastic_stiffness = 100.0;
+  float bending_stiffness = 0.0001;
+  float density = 0.1;
+  float damping = 0.01;
 };
 
 //********************************/
